requirements.txt
streamlit
pdfplumber
python-docx
pandas
pytesseract
Pillow
spacy
transformers
torch
openpyxl
